% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count_n_unique_emojis.R
\name{count_n_unique_emojis}
\alias{count_n_unique_emojis}
\title{Count the number of unique emojis}
\usage{
count_n_unique_emojis(lines)
}
\arguments{
\item{lines}{a text}
}
\description{
Count the number of unique emojis
}
\examples{
n <- count_n_unique_emojis(
  lines = readr::read_lines(get_splimata_path("example_1.md"))
)
testthat::expect_equal(2, n)
}
