% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_structured_chunk.R
\name{create_structured_chunk}
\alias{create_structured_chunk}
\title{Convert regular lines of one chunk to list}
\usage{
create_structured_chunk(chunk_lines)
}
\arguments{
\item{chunk_lines}{text of a chunk}
}
\description{
A chunk can be one of these, and one of these only:
}
\details{
\itemize{
\item regular lines of text
\item a heading
\item a table
}
}
\examples{
# Split text into chunks
text <- split_text(
  readr::read_lines(get_splimata_path("example_1.md"))
)
testthat::expect_true(is.list(text))
testthat::expect_equal(length(text), 4)

# Structure the chunks
testthat::expect_true(is.character(create_structured_chunk(text[[1]])))
testthat::expect_true(is.list(create_structured_chunk(text[[2]])))
testthat::expect_true(is.list(create_structured_chunk(text[[3]])))
testthat::expect_true(is.list(create_structured_chunk(text[[4]])))
}
