% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_structured_text.R
\name{create_structured_text}
\alias{create_structured_text}
\title{Convert regular lines of test to list}
\usage{
create_structured_text(lines)
}
\arguments{
\item{lines}{a text}
}
\description{
Convert regular lines of test to list
}
\examples{
text <- create_structured_text(
  readr::read_lines(get_splimata_path("example_1.md"))
)
testthat::expect_true(is.list(text))
}
