% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_text_per_language.R
\name{extract_text_per_language}
\alias{extract_text_per_language}
\title{Extract the languages}
\usage{
extract_text_per_language(lines)
}
\arguments{
\item{lines}{a text}
}
\description{
Extract the languages
}
\examples{
texts <- extract_text_per_language(
  lines = readr::read_lines(get_splimata_path("example_1.md"))
)
testthat::expect_true("en" \%in\% names(texts))
testthat::expect_true("sv" \%in\% names(texts))
testthat::expect_true(is.character(texts$en))
testthat::expect_true(is.character(texts$sv))
}
