% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_is_tab_line.R
\name{get_is_tab_line}
\alias{get_is_tab_line}
\title{Deterine per line if this is part of a tab yes/no}
\usage{
get_is_tab_line(lines)
}
\arguments{
\item{lines}{a text}
}
\description{
Deterine per line if this is part of a tab yes/no
}
\examples{
lines <- readr::read_lines(get_splimata_path("example_1.md"))
is_tab_line <- get_is_tab_line(lines)
testthat::expect_equal(length(lines), length(is_tab_line))
}
