% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_line_language_from_tab_chunk.R
\name{get_line_language_from_tab_chunk}
\alias{get_line_language_from_tab_chunk}
\title{From the text of a tab chunk, deduce which language each line belongs to}
\usage{
get_line_language_from_tab_chunk(chunk_lines)
}
\arguments{
\item{chunk_lines}{text of a chunk}
}
\description{
From the text of a tab chunk, deduce which language each line belongs to
}
\examples{
# Split text into chunks
text <- split_text(
  readr::read_lines(get_splimata_path("example_1.md"))
)
testthat::expect_true(is.list(text))
testthat::expect_equal(length(text), 4)

chunk_lines <- text[[4]]
line_language <- get_line_language_from_tab_chunk(chunk_lines)
testthat::expect_equal(length(chunk_lines), length(line_language))
}
