% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_unique_emojis.R
\name{get_unique_emojis}
\alias{get_unique_emojis}
\title{Get the unique emojis}
\usage{
get_unique_emojis(lines)
}
\arguments{
\item{lines}{a text}
}
\description{
Get the unique emojis
}
\examples{
unique_emojis <- get_unique_emojis(
  readr::read_lines(get_splimata_path("example_1.md"))
)
testthat::expect_equal(2, length(unique_emojis))
}
