% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_unique_language_codes.R
\name{get_unique_language_codes}
\alias{get_unique_language_codes}
\title{Get the unique emojis}
\usage{
get_unique_language_codes(lines)
}
\arguments{
\item{lines}{lines of text}
}
\description{
Get the unique emojis
}
