% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split_tabs.R
\name{split_tabs}
\alias{split_tabs}
\title{Split a MkDocs file based on language tabs.}
\usage{
split_tabs(input_file_name, output_file_prefix)
}
\arguments{
\item{input_file_name}{name of the file that has tabs with
multiple languages}

\item{output_file_prefix}{prefix of the created Markdown files,
e.g. \code{result}}
}
\value{
the names of the files created,
e.g. \code{c("result_en.md", "result_sv.md")}
}
\description{
Split a MkDocs file based on language tabs.
}
