% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split_text.R
\name{split_text}
\alias{split_text}
\title{Split text is different sections:
\itemize{
\item Headings
\item Regular text
\item Language tabs
}}
\usage{
split_text(lines)
}
\arguments{
\item{lines}{a text}
}
\description{
Split text is different sections:
\itemize{
\item Headings
\item Regular text
\item Language tabs
}
}
\examples{
text <- split_text(
  lines = readr::read_lines(get_splimata_path("example_1.md"))
)
testthat::expect_true(is.list(text))
}
