% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_splimata_path.R
\name{get_splimata_path}
\alias{get_splimata_path}
\title{Get the path to a \code{splimata} file}
\usage{
get_splimata_path(filename)
}
\arguments{
\item{filename}{name of the file in \code{inst/extdata}}
}
\description{
\itemize{
\item \href{https://uppsala-makerspace.github.io/loerdagskurser/volontaerer/bli_samordnare/}{example_1.md}
\item \href{https://uppsala-makerspace.github.io/loerdagskurser/plats/}{example_2.md}
}
}
